refactor: move plugin registration from observers.rs into lib.rs
- Move build_plugin logic from observers.rs into lib.rs as build_app,
  so plugin registration lives alongside the plugin struct definitions
- Both Plugin::build impls now call build_app directly instead of
  delegating to observers::build_plugin
- Make observer functions and run conditions pub(crate) so lib.rs can
  reference them as observers::fn_name
- Remove unused imports from observers.rs (PathBuf, monitors module,
  MonitorPlugin) now that it no longer owns plugin registration
- Update observers.rs module doc to reflect its reduced scope
- Change FRAME_EXTENT_COUNT constant type from usize to u32 to match
  its usage in the X11 property query (from style branch merge)
- Adjust x11_frame_extents.rs to cast FRAME_EXTENT_COUNT at the
  comparison site instead of the call site
- Follows new bevy-plugin-ownership style guide rule: plugin struct
  and impl Plugin live together in the module root

+++ src/constants.rs
@@ -35,7 +35,7 @@ pub(crate) const SUBCLASS_ID: usize = 1;
 // X11 frame extents (`_NET_FRAME_EXTENTS`: left, right, top, bottom)
 /// Number of values in `_NET_FRAME_EXTENTS` (left, right, top, bottom).
 #[cfg(all(target_os = "linux", feature = "workaround-winit-4445"))]
-pub(crate) const FRAME_EXTENT_COUNT: usize = 4;
+pub(crate) const FRAME_EXTENT_COUNT: u32 = 4;
 /// Index of the "top" extent in `_NET_FRAME_EXTENTS`.
 #[cfg(all(target_os = "linux", feature = "workaround-winit-4445"))]
 pub(crate) const FRAME_EXTENT_TOP_INDEX: usize = 2;
